Module Mdx.Toplevel

Toplevel phrases.

type t = {

vpad : int;

hpad : int;

pos : Lexing.position;

command : string list;

output : Output.t list;

}

The type for top-level phrases.

Pretty-printing

val dump : t Fmt.t

dump is the printer for dumping toplevel phrases. Useful for debugging.

val pp : t Fmt.t

pp is the pretty-printer for top-level phrases. pad is the size of the optionnalwhitespace left padding (by default is is 0).

val pp_command : t Fmt.t

pp_command is the pretty-printer for toplevel commands.

Parser

val of_lines : syntax:Syntax.t -> loc:Location.t -> string list -> t list

of_lines ~loc lines is the list of toplevel blocks from location loc. Return the vertical and horizontal whitespace padding as well.