Module Mdx__.Library

type t = {

base_name : string;

sub_lib : string option;

}

The type to represent dune libraries as referred to in '#require' statements or in dune files. I.e. lib.sub_lib is {base_name = "lib"; sub_lib = Some "sub_lib"}.

val equal : t -> t -> bool

val compare : t -> t -> int

val pp : t Fmt.t

val from_string : string -> (t, string) Result.result

from_string s returns the library represented by s or an error if s isn't a valid library.

module Set : sig ... end

Set of libraries