Module Mdx__Document

Lines

type syntax = Mdx.Syntax.t =

| Normal

| Cram

type line =

| Section of int * string

| Text of string

| Block of Mdx.Block.t

The type for the lines of a markdown or cram file.

val pp_line : ?⁠syntax:syntax -> line Fmt.t

pp_line is the pretty-printer for markdown or cram lines.

Document

type t = line list

The type for mdx documents.

val pp : ?⁠syntax:syntax -> t Fmt.t

pp is the pretty printer for mdx documents. Should be idempotent with of_string.

val to_string : t -> string

to_string t converts the document t to a string.